Optimize move-file method to use atomic move.
The move-file method did a lot of file operations that could mostly be
skipped on the "happy path". In some cases an atomic move is required
(e.g. moving a file to an inbox), debug logging will now show the
move-operation. Also made the "create directories" call synchronized to
prevent any race-conditions and log which directories are created.
Removed some pre Java 7 code in the file-delete method that is no longer
needed.